Create Python  to build the sensitivity.py
An oil company is considering converting its office building from a coal-burning furnace to
one that burns either fuel oil or natural gas. The cost of converting to fuel oil is estimated to be $80,000
initially; annual operating expenses are estimated to be $4,000 less than that experienced when using
the coal furnace. Approximately 140,000 Btus are produced per gallon of fuel oil; fuel oil is anticipated
to cost $2.20 per gallon.
The cost of converting to natural gas is estimated to be $60,000 initially; additionally, annual operating
and maintenance expenses are estimated to be $6,000 less than that for the coal-burning furnace.
Approximately 1,000 Btus are produced per cubic foot of natural gas; it is estimated that natural gas
will cost $0.04 per cubic foot.
A planning horizon of 20 years is to be used. Zero market values and a MARR of 10% per year are
appropriate. Perform a sensitivity analysis for the annual Btu requirement for the heating system.
(Hint: First calculate the breakeven number of Btus (in thousands). Then determine AWs if Btu
requirement varies over ±30% of the breakeven amount.)
Use functional symbols (factors) to show your logic and use Python or Excel to build the sensitivity
graph.